requirements.txt
fastapi
uvicorn
requests
PyGithub
python-dotenv
pytest